Format and clean code; Adjust benchmark test case
