+++ src/app/database/DataBase.cpp
@@ -1,9 +1,4 @@
 #include "DataBase.h"
-#include "leveldb/db.h"
-#include "leveldb/write_batch.h"
-#include "Config.h"
-#include "Log.h"
-#include "FileUtils.h"
 
 crust::Log *p_log = crust::Log::get_instance();
