+++ src/app/http/ApiHandler.cpp
@@ -1,6 +1,4 @@
 #include "ApiHandler.h"
-#include "sgx_tseal.h"
-#include <exception>
 
 namespace beast = boost::beast;         // from <boost/beast.hpp>
 namespace http = beast::http;           // from <boost/beast/http.hpp>
