+++ src/app/http/WebsocketClient.cpp
@@ -1,5 +1,4 @@
 #include "WebsocketClient.h"
-#include "Log.h"
 
 namespace beast = boost::beast; // from <boost/beast.hpp>
 namespace http = beast::http;   // from <boost/beast/http.hpp>
