+++ src/app/process/EntryNetwork.h
@@ -16,6 +16,7 @@
 #include "FormatUtils.h"
 #include "CrustStatus.h"
 #include "EnclaveData.h"
+#include "HttpClient.h"
 
 #define OPT_ISSET(x, y) x &y
 #define _rdrand64_step(x) ({ unsigned char err; asm volatile("rdrand %0; setc %1":"=r"(*x), "=qm"(err)); err; })
