+++ src/app/utils/SgxSupport.h
@@ -3,7 +3,28 @@
 
 #pragma once
 
+#include <limits.h>
+#include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+
+#include <sgx_edger8r.h>
+#include <sgx_uae_launch.h>
+#include <sgx_uae_epid.h>
+#include <sgx_uae_quote_ex.h>
 #include <sgx_urts.h>
+#include <sgx_urts.h>
+#include <sgx_capable.h>
+
+#include "Resource.h"
+
+#ifdef _WIN32
+#include <Windows.h>
+#else
+#include <dlfcn.h>
+#endif
 
 #define SGX_SUPPORT_UNKNOWN 0x00000000
 #define SGX_SUPPORT_NO 0x08000000
