+++ src/enclave/persistence/Persistence.cpp
@@ -1,7 +1,4 @@
 #include "Persistence.h"
-#include "EUtils.h"
-#include "sgx_tseal.h"
-#include "EJson.h"
 
 using namespace std;
