+++ src/enclave/utils/EUtils.cpp
@@ -1,5 +1,5 @@
 #include "EUtils.h"
-#include "EJson.h"
+#include "Json.h"
 
 using namespace std;
