+++ src/enclave/utils/Json.h
@@ -11,9 +11,15 @@
 #include <map>
 #include <type_traits>
 #include <initializer_list>
+#ifdef _CRUST_RESOURCE_H_
 #include <ostream>
 #include <iostream>
-
+#include <string.h>
+#include "FormatUtils.h"
+#else
+#include "EUtils.h"
+#endif
+#define HASH_TAG "$&JT&$"
 #define JSON_NL "%$&nl&$%"
 
 namespace json
@@ -28,6 +34,8 @@ using std::is_integral;
 using std::is_same;
 using std::map;
 using std::string;
+
+const uint32_t _hash_length = 32;
 
 namespace
 {
@@ -122,6 +130,33 @@ string json_escape(const string &str)
     str_replace(output, JSON_NL, "\n        ");
     return std::move(output);
 }
+std::string _hexstring(const void *vsrc, size_t len)
+{
+    size_t i;
+    const uint8_t *src = (const uint8_t *)vsrc;
+    char *hex_buffer = (char*)malloc(len * 2);
+    if (hex_buffer == NULL)
+    {
+        return "";
+    }
+    memset(hex_buffer, 0, len * 2);
+    char *bp;
+    const char _hextable[] = "0123456789abcdef";
+
+    for (i = 0, bp = hex_buffer; i < len; ++i)
+    {
+        *bp = _hextable[src[i] >> 4];
+        ++bp;
+        *bp = _hextable[src[i] & 0xf];
+        ++bp;
+    }
+
+    std::string ans(hex_buffer, len * 2);
+
+    free(hex_buffer);
+
+    return ans;
+}
 } // namespace
 
 class JSON
@@ -134,6 +169,7 @@ class JSON
         BackingData() : Int(0) {}
 
         deque<JSON> *List;
+        uint8_t *HashList;
         map<string, JSON> *Map;
         string *String;
         double Float;
@@ -147,6 +183,7 @@ public:
         Null,
         Object,
         Array,
+        Hash,
         String,
         Floating,
         Integral,
@@ -222,6 +259,10 @@ public:
                 new deque<JSON>(other.Internal.List->begin(),
                                 other.Internal.List->end());
             break;
+        case Class::Hash:
+            Internal.HashList = new uint8_t[_hash_length];
+            memcpy(Internal.HashList, other.Internal.HashList, _hash_length);
+            break;
         case Class::String:
             Internal.String =
                 new string(*other.Internal.String);
@@ -247,6 +288,10 @@ public:
                 new deque<JSON>(other.Internal.List->begin(),
                                 other.Internal.List->end());
             break;
+        case Class::Hash:
+            Internal.HashList = new uint8_t[_hash_length];
+            memcpy(Internal.HashList, other.Internal.HashList, _hash_length);
+            break;
         case Class::String:
             Internal.String =
                 new string(*other.Internal.String);
@@ -264,6 +309,9 @@ public:
         {
         case Class::Array:
             delete Internal.List;
+            break;
+        case Class::Hash:
+            delete[] Internal.HashList;
             break;
         case Class::Object:
             delete Internal.Map;
@@ -297,6 +345,7 @@ public:
     }
 
     static JSON Load(const string &);
+    static JSON Load(const uint8_t *p_data, size_t data_size);
 
     template <typename T>
     void append(T arg)
@@ -324,7 +373,7 @@ public:
     typename enable_if<is_integral<T>::value && !is_same<T, bool>::value, JSON &>::type operator=(T i)
     {
         SetType(Class::Integral);
-        Internal.Int = i;
+        Internal.Int = (long)i;
         return *this;
     }
 
@@ -344,6 +393,15 @@ public:
         return *this;
     }
 
+    template <typename T>
+    typename enable_if<is_same<T, uint8_t*>::value, JSON &>::type operator=(T v)
+    {
+        SetType(Class::Hash);
+        Internal.HashList = new uint8_t[_hash_length];
+        memcpy(Internal.HashList, v, _hash_length);
+        return *this;
+    }
+
     JSON &operator[](const string &key)
     {
         SetType(Class::Object);
@@ -356,6 +414,25 @@ public:
         if (index >= Internal.List->size())
             Internal.List->resize(index + 1);
         return Internal.List->operator[](index);
+    }
+
+    char get_char(size_t index)
+    {
+        if (Type != Class::String || index >= Internal.String->size())
+            return '\0';
+
+        return Internal.String->operator[](index);
+    }
+
+    void set_char(size_t index, char c)
+    {
+        if (Type != Class::String || index > Internal.String->size())
+            return;
+
+        if (index == Internal.String->size())
+            Internal.String->push_back(c);
+        else
+            Internal.String->operator[](index) = c;
     }
 
     JSON &at(const string &key)
@@ -378,10 +455,12 @@ public:
         return Internal.List->at(index);
     }
 
-    int length() const
+    long length() const
     {
         if (Type == Class::Array)
-            return (int)Internal.List->size();
+            return (long)Internal.List->size();
+        else if (Type == Class::Hash)
+            return _hash_length;
         else
             return -1;
     }
@@ -393,12 +472,14 @@ public:
         return false;
     }
 
-    int size() const
+    long size() const
     {
         if (Type == Class::Object)
-            return (int)Internal.Map->size();
+            return (long)Internal.Map->size();
         else if (Type == Class::Array)
-            return (int)Internal.List->size();
+            return (long)Internal.List->size();
+        else if (Type == Class::Hash)
+            return _hash_length;
         else
             return -1;
     }
@@ -416,7 +497,25 @@ public:
     string ToString(bool &ok) const
     {
         ok = (Type == Class::String);
-        return ok ? std::move(json_escape(*Internal.String)) : string("");
+        if (Type == Class::String)
+            return std::move(json_escape(*Internal.String));
+        else if (Type == Class::Hash)
+            return _hexstring(Internal.HashList, _hash_length);
+        else if (Type == Class::Integral)
+            return std::to_string(Internal.Int);
+        else
+            return string("");
+    }
+
+    uint8_t *ToBytes()
+    {
+        bool b;
+        return ToBytes(b);
+    }
+    uint8_t *ToBytes(bool &ok)
+    {
+        ok = (Type == Class::Hash);
+        return ok ? Internal.HashList : NULL;
     }
 
     double ToFloat() const
@@ -518,6 +617,10 @@ public:
             s += "\n" + pad.erase(0, 2) + "]";
             return s;
         }
+        case Class::Hash:
+        {
+            return "\"" HASH_TAG + _hexstring(Internal.HashList, _hash_length) + "\"";
+        }
         case Class::String:
             return "\"" + json_escape_pad(*Internal.String, pad) + "\"";
         case Class::Floating:
@@ -532,7 +635,7 @@ public:
         return "";
     }
 
-    friend std::ostream &operator<<(std::ostream &, const JSON &);
+    //friend std::ostream &operator<<(std::ostream &, const JSON &);
 
 private:
     void SetType(Class type)
@@ -552,6 +655,9 @@ private:
             break;
         case Class::Array:
             Internal.List = new deque<JSON>();
+            break;
+        case Class::Hash:
+            Internal.HashList = new uint8_t[_hash_length];
             break;
         case Class::String:
             Internal.String = new string();
@@ -585,6 +691,9 @@ private:
         case Class::Array:
             delete Internal.List;
             break;
+        case Class::Hash:
+            delete[] Internal.HashList;
+            break;
         case Class::String:
             delete Internal.String;
             break;
@@ -614,19 +723,29 @@ JSON Object()
     return std::move(JSON::Make(JSON::Class::Object));
 }
 
+/*
 std::ostream &operator<<(std::ostream &os, const JSON &json)
 {
     os << json.dump();
     return os;
 }
+*/
 
 namespace
 {
 JSON parse_next(const string &, size_t &);
 
+JSON parse_next(const uint8_t *p_data, size_t &offset);
+
 void consume_ws(const string &str, size_t &offset)
 {
     while (isspace(str[offset]))
+        ++offset;
+}
+
+void consume_ws(const uint8_t *p_data, size_t &offset)
+{
+    while (isspace(p_data[offset]))
         ++offset;
 }
 
@@ -648,7 +767,11 @@ JSON parse_object(const string &str, size_t &offset)
         consume_ws(str, offset);
         if (str[offset] != ':')
         {
+#ifdef _CRUST_RESOURCE_H_
             std::cerr << "Error: Object: Expected colon, found '" << str[offset] << "'\n";
+#else
+            log_err("Error: Object: Expected colon, found %c\n", str[offset]);
+#endif
             break;
         }
         consume_ws(str, ++offset);
@@ -668,7 +791,65 @@ JSON parse_object(const string &str, size_t &offset)
         }
         else
         {
+#ifdef _CRUST_RESOURCE_H_
             std::cerr << "ERROR: Object: Expected comma, found '" << str[offset] << "'\n";
+#else
+            log_err("Error: Object: Expected comma, found %c\n", str[offset]);
+#endif
+            break;
+        }
+    }
+
+    return std::move(Object);
+}
+
+JSON parse_object(const uint8_t *p_data, size_t &offset)
+{
+    JSON Object = JSON::Make(JSON::Class::Object);
+
+    ++offset;
+    consume_ws(p_data, offset);
+    if (p_data[offset] == '}')
+    {
+        ++offset;
+        return std::move(Object);
+    }
+
+    while (true)
+    {
+        JSON Key = parse_next(p_data, offset);
+        consume_ws(p_data, offset);
+        if (p_data[offset] != ':')
+        {
+#ifdef _CRUST_RESOURCE_H_
+            std::cerr << "Error: Object: Expected colon, found '" << p_data[offset] << "'\n";
+#else
+            log_err("Error: Object: Expected colon, found %c\n", p_data[offset]);
+#endif
+            break;
+        }
+        consume_ws(p_data, ++offset);
+        JSON Value = parse_next(p_data, offset);
+        Object[Key.ToString()] = Value;
+
+        consume_ws(p_data, offset);
+        if (p_data[offset] == ',')
+        {
+            ++offset;
+            continue;
+        }
+        else if (p_data[offset] == '}')
+        {
+            ++offset;
+            break;
+        }
+        else
+        {
+#ifdef _CRUST_RESOURCE_H_
+            std::cerr << "ERROR: Object: Expected comma, found '" << p_data[offset] << "'\n";
+#else
+            log_err("Error: Object: Expected comma, found %c\n", p_data[offset]);
+#endif
             break;
         }
     }
@@ -706,7 +887,53 @@ JSON parse_array(const string &str, size_t &offset)
         }
         else
         {
+#ifdef _CRUST_RESOURCE_H_
             std::cerr << "ERROR: Array: Expected ',' or ']', found '" << str[offset] << "'\n";
+#else
+            log_err("Error: Array: Expected  ',' or ']', found %c\n", str[offset]);
+#endif
+            return std::move(JSON::Make(JSON::Class::Array));
+        }
+    }
+
+    return std::move(Array);
+}
+
+JSON parse_array(const uint8_t *p_data, size_t &offset)
+{
+    JSON Array = JSON::Make(JSON::Class::Array);
+    unsigned index = 0;
+
+    ++offset;
+    consume_ws(p_data, offset);
+    if (p_data[offset] == ']')
+    {
+        ++offset;
+        return std::move(Array);
+    }
+
+    while (true)
+    {
+        Array[index++] = parse_next(p_data, offset);
+        consume_ws(p_data, offset);
+
+        if (p_data[offset] == ',')
+        {
+            ++offset;
+            continue;
+        }
+        else if (p_data[offset] == ']')
+        {
+            ++offset;
+            break;
+        }
+        else
+        {
+#ifdef _CRUST_RESOURCE_H_
+            std::cerr << "ERROR: Array: Expected ',' or ']', found '" << p_data[offset] << "'\n";
+#else
+            log_err("Error: Array: Expected  ',' or ']', found %c\n", p_data[offset]);
+#endif
             return std::move(JSON::Make(JSON::Class::Array));
         }
     }
@@ -716,7 +943,7 @@ JSON parse_array(const string &str, size_t &offset)
 
 JSON parse_string(const string &str, size_t &offset)
 {
-    JSON String;
+    JSON ans;
     string val;
     for (char c = str[++offset]; c != '\"'; c = str[++offset])
     {
@@ -758,7 +985,11 @@ JSON parse_string(const string &str, size_t &offset)
                         val += c;
                     else
                     {
+#ifdef _CRUST_RESOURCE_H_
                         std::cerr << "ERROR: String: Expected hex character in unicode escape, found '" << c << "'\n";
+#else
+                        log_err("Error: String: Expected hex character in unicode escape, found %c\n", c);
+#endif
                         return std::move(JSON::Make(JSON::Class::String));
                     }
                 }
@@ -774,8 +1005,98 @@ JSON parse_string(const string &str, size_t &offset)
             val += c;
     }
     ++offset;
-    String = val;
-    return std::move(String);
+    if (memcmp(val.c_str(), HASH_TAG, strlen(HASH_TAG)) == 0)
+    {
+        val = val.substr(strlen(HASH_TAG), val.size());
+        uint8_t *p_hash = hex_string_to_bytes(val.c_str(), val.size());
+        if (p_hash != NULL)
+        {
+            ans = p_hash;
+            free(p_hash);
+            return std::move(ans);
+        }
+    }
+    ans = val;
+    return std::move(ans);
+}
+
+JSON parse_string(const uint8_t *p_data, size_t &offset)
+{
+    JSON ans;
+    string val;
+    for (char c = p_data[++offset]; c != '\"'; c = p_data[++offset])
+    {
+        if (c == '\\')
+        {
+            switch (p_data[++offset])
+            {
+            case '\"':
+                val += '\"';
+                break;
+            case '\\':
+                val += '\\';
+                break;
+            case '/':
+                val += '/';
+                break;
+            case 'b':
+                val += '\b';
+                break;
+            case 'f':
+                val += '\f';
+                break;
+            case 'n':
+                val += '\n';
+                break;
+            case 'r':
+                val += '\r';
+                break;
+            case 't':
+                val += '\t';
+                break;
+            case 'u':
+            {
+                val += "\\u";
+                for (unsigned i = 1; i <= 4; ++i)
+                {
+                    c = p_data[offset + i];
+                    if ((c >= '0' && c <= '9') || (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F'))
+                        val += c;
+                    else
+                    {
+#ifdef _CRUST_RESOURCE_H_
+                        std::cerr << "ERROR: String: Expected hex character in unicode escape, found '" << c << "'\n";
+#else
+                        log_err("Error: String: Expected hex character in unicode escape, found %c\n", c);
+#endif
+                        return std::move(JSON::Make(JSON::Class::String));
+                    }
+                }
+                offset += 4;
+            }
+            break;
+            default:
+                val += '\\';
+                break;
+            }
+        }
+        else
+            val += c;
+    }
+    ++offset;
+    if (memcmp(val.c_str(), HASH_TAG, strlen(HASH_TAG)) == 0)
+    {
+        val = val.substr(strlen(HASH_TAG), val.size());
+        uint8_t *p_hash = hex_string_to_bytes(val.c_str(), val.size());
+        if (p_hash != NULL)
+        {
+            ans = p_hash;
+            free(p_hash);
+            return std::move(ans);
+        }
+    }
+    ans = val;
+    return std::move(ans);
 }
 
 JSON parse_number(const string &str, size_t &offset)
@@ -813,7 +1134,11 @@ JSON parse_number(const string &str, size_t &offset)
                 exp_str += c;
             else if (!isspace(c) && c != ',' && c != ']' && c != '}')
             {
+#ifdef _CRUST_RESOURCE_H_
                 std::cerr << "ERROR: Number: Expected a number for exponent, found '" << c << "'\n";
+#else
+                log_err("Error: Number: Expected a number for exponent, found %c\n", c);
+#endif
                 return std::move(JSON::Make(JSON::Class::Null));
             }
             else
@@ -823,7 +1148,81 @@ JSON parse_number(const string &str, size_t &offset)
     }
     else if (!isspace(c) && c != ',' && c != ']' && c != '}')
     {
+#ifdef _CRUST_RESOURCE_H_
         std::cerr << "ERROR: Number: unexpected character '" << c << "'\n";
+#else
+        log_err("Error: Number: unexpected character %c\n", c);
+#endif
+        return std::move(JSON::Make(JSON::Class::Null));
+    }
+    --offset;
+
+    if (isDouble)
+        Number = std::stod(val) * std::pow(10, exp);
+    else
+    {
+        if (!exp_str.empty())
+            Number = (double)std::stol(val) * (double)std::pow(10, exp);
+        else
+            Number = std::stol(val);
+    }
+    return std::move(Number);
+}
+
+JSON parse_number(const uint8_t *p_data, size_t &offset)
+{
+    JSON Number;
+    string val, exp_str;
+    char c;
+    bool isDouble = false;
+    long exp = 0;
+    while (true)
+    {
+        c = p_data[offset++];
+        if ((c == '-') || (c >= '0' && c <= '9'))
+            val += c;
+        else if (c == '.')
+        {
+            val += c;
+            isDouble = true;
+        }
+        else
+            break;
+    }
+    if (c == 'E' || c == 'e')
+    {
+        c = p_data[offset++];
+        if (c == '-')
+        {
+            ++offset;
+            exp_str += '-';
+        }
+        while (true)
+        {
+            c = p_data[offset++];
+            if (c >= '0' && c <= '9')
+                exp_str += c;
+            else if (!isspace(c) && c != ',' && c != ']' && c != '}')
+            {
+#ifdef _CRUST_RESOURCE_H_
+                std::cerr << "ERROR: Number: Expected a number for exponent, found '" << c << "'\n";
+#else
+                log_err("Error: Number: Expected a number for exponent, found %c\n", c);
+#endif
+                return std::move(JSON::Make(JSON::Class::Null));
+            }
+            else
+                break;
+        }
+        exp = std::stol(exp_str);
+    }
+    else if (!isspace(c) && c != ',' && c != ']' && c != '}')
+    {
+#ifdef _CRUST_RESOURCE_H_
+        std::cerr << "ERROR: Number: unexpected character '" << c << "'\n";
+#else
+        log_err("Error: Number: unexpected character %c\n", c);
+#endif
         return std::move(JSON::Make(JSON::Class::Null));
     }
     --offset;
@@ -849,7 +1248,31 @@ JSON parse_bool(const string &str, size_t &offset)
         Bool = false;
     else
     {
+#ifdef _CRUST_RESOURCE_H_
         std::cerr << "ERROR: Bool: Expected 'true' or 'false', found '" << str.substr(offset, 5) << "'\n";
+#else
+        log_err("Error: Bool: Expected 'true' or 'false', found %s\n", str.substr(offset, 5));
+#endif
+        return std::move(JSON::Make(JSON::Class::Null));
+    }
+    offset += (Bool.ToBool() ? 4 : 5);
+    return std::move(Bool);
+}
+
+JSON parse_bool(const uint8_t *p_data, size_t &offset)
+{
+    JSON Bool;
+    if (memcmp(p_data + offset, "true", 4) == 0)
+        Bool = true;
+    else if (memcmp(p_data + offset, "false", 5) == 0)
+        Bool = false;
+    else
+    {
+#ifdef _CRUST_RESOURCE_H_
+        std::cerr << "ERROR: Bool: Expected 'true' or 'false', found '" << hexstring_safe(p_data + offset, 5) << "'\n";
+#else
+        log_err("Error: Bool: Expected 'true' or 'false', found error\n");
+#endif
         return std::move(JSON::Make(JSON::Class::Null));
     }
     offset += (Bool.ToBool() ? 4 : 5);
@@ -861,7 +1284,27 @@ JSON parse_null(const string &str, size_t &offset)
     JSON Null;
     if (str.substr(offset, 4) != "null")
     {
+#ifdef _CRUST_RESOURCE_H_
         std::cerr << "ERROR: Null: Expected 'null', found '" << str.substr(offset, 4) << "'\n";
+#else
+        log_err("Error: Null: Expected 'null', found %s\n", str.substr(offset, 4));
+#endif
+        return std::move(JSON::Make(JSON::Class::Null));
+    }
+    offset += 4;
+    return std::move(Null);
+}
+
+JSON parse_null(const uint8_t *p_data, size_t &offset)
+{
+    JSON Null;
+    if (memcmp(p_data + offset, "null", 4) == 0)
+    {
+#ifdef _CRUST_RESOURCE_H_
+        std::cerr << "ERROR: Null: Expected 'null', found '" << hexstring_safe(p_data + offset, 4) << "'\n";
+#else
+        log_err("Error: Null: Expected 'null', found error\n");
+#endif
         return std::move(JSON::Make(JSON::Class::Null));
     }
     offset += 4;
@@ -890,7 +1333,41 @@ JSON parse_next(const string &str, size_t &offset)
         if ((value <= '9' && value >= '0') || value == '-')
             return std::move(parse_number(str, offset));
     }
+#ifdef _CRUST_RESOURCE_H_
     std::cerr << "ERROR: Parse: Unknown starting character '" << value << "'\n";
+#else
+    log_err("Error: Parse: Unknown starting character %c\n", value);
+#endif
+    return JSON();
+}
+
+JSON parse_next(const uint8_t *p_data, size_t &offset)
+{
+    char value;
+    consume_ws(p_data, offset);
+    value = p_data[offset];
+    switch (value)
+    {
+    case '[':
+        return std::move(parse_array(p_data, offset));
+    case '{':
+        return std::move(parse_object(p_data, offset));
+    case '\"':
+        return std::move(parse_string(p_data, offset));
+    case 't':
+    case 'f':
+        return std::move(parse_bool(p_data, offset));
+    case 'n':
+        return std::move(parse_null(p_data, offset));
+    default:
+        if ((value <= '9' && value >= '0') || value == '-')
+            return std::move(parse_number(p_data, offset));
+    }
+#ifdef _CRUST_RESOURCE_H_
+    std::cerr << "ERROR: Parse: Unknown starting character '" << value << "'\n";
+#else
+    log_err("Error: Parse: Unknown starting character %c\n", value);
+#endif
     return JSON();
 }
 } // namespace
@@ -900,6 +1377,13 @@ JSON JSON::Load(const string &str)
     if (str.size() == 0) return json::JSON();
     size_t offset = 0;
     return std::move(parse_next(str, offset));
+}
+
+JSON JSON::Load(const uint8_t *p_data, size_t data_size)
+{
+    if (data_size == 0) return json::JSON();
+    size_t offset = 0;
+    return std::move(parse_next(p_data, offset));
 }
 
 } // End Namespace json
