+++ test/src/app/process/SrdTest.h
@@ -7,7 +7,6 @@
 #include <string>
 #include <vector>
 #include <unordered_set>
-#include "Json.hpp"
 #include "Config.h"
 #include "FileUtils.h"
 #include "FormatUtils.h"
