+++ test/src/enclave/report/ReportTest.cpp
@@ -1,5 +1,4 @@
 #include "ReportTest.h"
-#include "EJson.h"
 
 
 std::string g_work_report_test;
